+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### [1.1.2] - 2022-03-31
+- Add default commands to the manifest (package.json)
+- Add this commands to the merged configuration
+
 ### [1.1.1] - 2022-03-28
 - Fix a bug in the configuration merge